# Skip ASCII check until the non ASCII byte is found