@lpinca 2021-08-12
Master branch

```
$ ./build/benchmarks/benchmark -P validate_utf8+fallback -F ~/data/data.txt 
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
testcases: 1
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 10485760, iterations: 2000, dataset: /Users/luigi/data/data.txt
   1.064 GB/s (0.6 %)    0.997 Gc/s     1.07 byte/char
```

This PR

```
$ ./build/benchmarks/benchmark -P validate_utf8+fallback -F ~/data/data.txt 
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
testcases: 1
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 10485760, iterations: 2000, dataset: /Users/luigi/data/data.txt
   2.542 GB/s (1.5 %)    2.383 Gc/s     1.07 byte/char
```

The dataset is a 10 MiB file made of `aaaaaaaaaaaaaa©` chunks.

Will run more benchmarks with real world dataset later.