@lemire 2021-08-12
I am concerned about such a synthetic dataset.

Would you try again with the files from this repository ?

https://github.com/lemire/unicode_lipsum