@lpinca 2021-08-12
Wikipedia Japanese main page:

Master branch

```
$ ./build/benchmarks/benchmark -P validate_utf8+fallback -F ~/data/wikipedia.ja.html 
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
testcases: 1
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 98576, iterations: 2000, dataset: /Users/luigi/data/wikipedia.ja.html
   2.204 GB/s (1.9 %)    1.938 Gc/s     1.14 byte/char
```

This PR

```
$ ./build/benchmarks/benchmark -P validate_utf8+fallback -F ~/data/wikipedia.ja.html 
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
testcases: 1
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 98576, iterations: 2000, dataset: /Users/luigi/data/wikipedia.ja.html
   3.377 GB/s (2.6 %)    2.969 Gc/s     1.14 byte/char
```