@lpinca 2021-08-12
From a quick look almost all files in that repository with the exception of `Russian-Lipsum.utf8.txt` and `Latin-Lipsum.utf8.txt` are made exclusively of multi bytes characters so it doesn't make much sense but will do.