@lpinca 2021-08-12
Master branch

```
$ ./build/benchmarks/benchmark -P validate_utf8+fallback -F ~/unicode_lipsum/wikipedia_mars/*.html
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
testcases: 18
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 954430, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/arabic.html
   1.588 GB/s (2.1 %)    1.388 Gc/s     1.14 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 382079, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/chinese.html
   2.031 GB/s (2.3 %)    1.787 Gc/s     1.14 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 368442, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/czech.html
   2.346 GB/s (2.7 %)    2.287 Gc/s     1.03 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 1005060, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/english.html
  12.123 GB/s (2.2 %)   12.088 Gc/s     1.00 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 192461, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/esperanto.html
   4.831 GB/s (3.6 %)    4.757 Gc/s     1.02 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 1032638, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/french.html
   3.584 GB/s (2.9 %)    3.539 Gc/s     1.01 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 397376, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/german.html
   4.850 GB/s (2.9 %)    4.793 Gc/s     1.01 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 326722, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/greek.html
   1.693 GB/s (2.7 %)    1.482 Gc/s     1.14 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 327412, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/hebrew.html
   1.320 GB/s (2.4 %)    1.136 Gc/s     1.16 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 712465, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/hindi.html
   1.585 GB/s (1.9 %)    1.301 Gc/s     1.22 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 304786, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/japanese.html
   1.790 GB/s (2.3 %)    1.509 Gc/s     1.19 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 193001, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/korean.html
   1.559 GB/s (2.4 %)    1.346 Gc/s     1.16 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 293677, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/persan.html
   1.608 GB/s (2.8 %)    1.425 Gc/s     1.13 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 692409, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/portuguese.html
   4.270 GB/s (2.6 %)    4.225 Gc/s     1.01 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 713817, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/russian.html
   1.421 GB/s (2.2 %)    1.217 Gc/s     1.17 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 1088085, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/thai.html
   1.972 GB/s (2.0 %)    1.618 Gc/s     1.22 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 387007, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/turkish.html
   2.319 GB/s (2.8 %)    2.259 Gc/s     1.03 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 674255, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/vietnamese.html
   1.614 GB/s (2.2 %)    1.523 Gc/s     1.06 byte/char
```

This PR

```
$ ./build/benchmarks/benchmark -P validate_utf8+fallback -F ~/unicode_lipsum/wikipedia_mars/*.html
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
testcases: 18
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 954430, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/arabic.html
   1.763 GB/s (2.4 %)    1.541 Gc/s     1.14 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 382079, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/chinese.html
   2.908 GB/s (2.7 %)    2.559 Gc/s     1.14 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 368442, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/czech.html
   2.869 GB/s (5.5 %)    2.797 Gc/s     1.03 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 1005060, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/english.html
  16.369 GB/s (3.1 %)   16.322 Gc/s     1.00 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 192461, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/esperanto.html
   6.678 GB/s (3.3 %)    6.575 Gc/s     1.02 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 1032638, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/french.html
   4.593 GB/s (3.4 %)    4.535 Gc/s     1.01 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 397376, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/german.html
   6.486 GB/s (2.6 %)    6.411 Gc/s     1.01 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 326722, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/greek.html
   1.907 GB/s (2.9 %)    1.670 Gc/s     1.14 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 327412, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/hebrew.html
   1.457 GB/s (2.8 %)    1.254 Gc/s     1.16 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 712465, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/hindi.html
   2.229 GB/s (2.4 %)    1.830 Gc/s     1.22 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 304786, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/japanese.html
   2.589 GB/s (2.7 %)    2.183 Gc/s     1.19 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 193001, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/korean.html
   2.118 GB/s (3.4 %)    1.829 Gc/s     1.16 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 293677, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/persan.html
   1.803 GB/s (3.0 %)    1.598 Gc/s     1.13 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 692409, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/portuguese.html
   5.480 GB/s (3.1 %)    5.422 Gc/s     1.01 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 713817, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/russian.html
   1.612 GB/s (3.4 %)    1.380 Gc/s     1.17 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 1088085, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/thai.html
   2.971 GB/s (2.1 %)    2.437 Gc/s     1.22 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 387007, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/turkish.html
   2.819 GB/s (5.3 %)    2.746 Gc/s     1.03 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 674255, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/wikipedia_mars/vietnamese.html
   2.034 GB/s (4.0 %)    1.920 Gc/s     1.06 byte/char
```