@lpinca 2021-08-12
Master branch

```
$ ./build/benchmarks/benchmark -P validate_utf8+fallback -F ~/unicode_lipsum/lipsum/*.utf8.txt
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
testcases: 9
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 81685, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/lipsum/Arabic-Lipsum.utf8.txt
   0.657 GB/s (3.0 %)    0.368 Gc/s     1.78 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 69840, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/lipsum/Chinese-Lipsum.utf8.txt
   0.778 GB/s (1.6 %)    0.261 Gc/s     2.98 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 65542, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/lipsum/Emoji-Lipsum.utf8.txt
   0.985 GB/s (1.7 %)    0.246 Gc/s     4.00 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 66495, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/lipsum/Hebrew-Lipsum.utf8.txt
   0.646 GB/s (3.3 %)    0.362 Gc/s     1.78 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 87997, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/lipsum/Hindi-Lipsum.utf8.txt
   0.632 GB/s (1.9 %)    0.235 Gc/s     2.69 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 67808, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/lipsum/Japanese-Lipsum.utf8.txt
   0.748 GB/s (6.2 %)    0.258 Gc/s     2.90 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 66600, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/lipsum/Korean-Lipsum.utf8.txt
   0.780 GB/s (4.5 %)    0.318 Gc/s     2.45 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 86940, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt
  16.576 GB/s (7.9 %)   16.576 Gc/s     1.00 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 104770, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/lipsum/Russian-Lipsum.utf8.txt
   0.591 GB/s (2.4 %)    0.327 Gc/s     1.81 byte/char
```

This PR

```
$ ./build/benchmarks/benchmark -P validate_utf8+fallback -F ~/unicode_lipsum/lipsum/*.utf8.txt
We define the number of bytes to be the number of *input* bytes.
We define a 'char' to be a code point (between 1 and 4 bytes).
===========================
testcases: 9
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 81685, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/lipsum/Arabic-Lipsum.utf8.txt
   0.642 GB/s (3.4 %)    0.360 Gc/s     1.78 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 69840, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/lipsum/Chinese-Lipsum.utf8.txt
   1.256 GB/s (2.3 %)    0.422 Gc/s     2.98 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 65542, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/lipsum/Emoji-Lipsum.utf8.txt
   1.196 GB/s (1.4 %)    0.299 Gc/s     4.00 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 66495, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/lipsum/Hebrew-Lipsum.utf8.txt
   0.626 GB/s (3.0 %)    0.351 Gc/s     1.78 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 87997, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/lipsum/Hindi-Lipsum.utf8.txt
   0.928 GB/s (2.3 %)    0.345 Gc/s     2.69 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 67808, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/lipsum/Japanese-Lipsum.utf8.txt
   1.198 GB/s (1.6 %)    0.413 Gc/s     2.90 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 66600, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/lipsum/Korean-Lipsum.utf8.txt
   1.150 GB/s (3.0 %)    0.469 Gc/s     2.45 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 86940, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/lipsum/Latin-Lipsum.utf8.txt
  22.083 GB/s (1.1 %)   22.083 Gc/s     1.00 byte/char 
input detected as UTF8
current system detected as haswell
===========================
validate_utf8+fallback, input size: 104770, iterations: 2000, dataset: /Users/luigi/unicode_lipsum/lipsum/Russian-Lipsum.utf8.txt
   0.641 GB/s (2.4 %)    0.355 Gc/s     1.81 byte/char
```